Consolidate R wrapper modules into _r_wrapper
Move CircE and skew-normal wrapper functions into `_r_wrapper`,
remove the old split modules, and update imports and tests to use
the unified internal API.
