+++ test/spi/test_r_wrapper.py
@@ -7,6 +7,7 @@ They are skipped if rpy2 is not installed.
 
 import pytest
 
+
 # === Pure-Python tests (no R required) ===
 
 
@@ -32,48 +33,37 @@ def test_ver_avoids_lexicographic_pitfall():
 
 
 class TestRWrapper:
-    """Test the R wrapper functionality."""
+    """Test the R wrapper session management."""
 
-    def test_initialize_r_session(self):
-        """Test R session initialization."""
-        from soundscapy.r_wrapper._r_wrapper import _state, initialize_r_session
+    def test_get_r_session_returns_active(self):
+        """get_r_session() should return an active, initialised session."""
+        from soundscapy.r_wrapper._r_wrapper import get_r_session
 
-        initialize_r_session()
-        assert _state.active, "R session should be active"
-        assert _state.circe_sourced, "CircE should be loaded from embedded scripts"
+        r = get_r_session()
+        assert r.active, "R session should be active after get_r_session()"
+        assert r.sn is not None, "sn package should be loaded"
+        assert r.base is not None, "base package should be loaded"
 
     def test_reset_r_session(self):
-        """Test R session package unloading."""
-        from soundscapy.r_wrapper._r_wrapper import reset_r_session
+        """reset_r_session() should succeed and deactivate the session."""
+        import soundscapy.r_wrapper._r_wrapper as rw
 
-        res = reset_r_session()
-        assert res, "R session packages should be unloaded successfully"
+        res = rw.reset_r_session()
+        assert res, "reset_r_session() should return True on success"
+        # Access _state via module to get the newly-bound object after reset.
+        assert not rw._state.active, "session should be inactive after reset"
 
     def test_r_session_reinitialization(self):
-        """Test that the R session can be reinitialized after reset."""
-        from soundscapy.r_wrapper._r_wrapper import (
-            _state,
-            initialize_r_session,
-            reset_r_session,
+        """The session can be reset and then re-initialised transparently."""
+        import soundscapy.r_wrapper._r_wrapper as rw
+
+        rw.get_r_session()
+        assert rw._state.active
+
+        rw.reset_r_session()
+        assert not rw._state.active
+
+        rw.get_r_session()
+        assert rw._state.active, (
+            "session should be active again after re-initialisation"
         )
-
-        initialize_r_session()
-        assert _state.active, "R session should be initialized"
-
-        reset_res = reset_r_session()
-        assert reset_res, "R session packages should be unloaded successfully"
-
-        initialize_r_session()
-        assert _state.active, "R session should be reinitialized"
-
-    def test_check_sn_package(self):
-        """Test that the R 'sn' package is available when R deps are installed."""
-        import soundscapy.r_wrapper as sspyr
-
-        sspyr._r_wrapper.check_sn_package()
-
-    def test_check_circe_package(self):
-        """Test that the embedded CircE scripts are available when R deps are installed."""
-        import soundscapy.r_wrapper as sspyr
-
-        sspyr._r_wrapper.check_circe_package()
